requirements.txt
# Bot framework
python-telegram-bot>=21.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Development tools (optional)
pyright>=1.1.0
